Split IO libs even more
Allowing not having to link various libs to various executables.
