+++ src/UnitTest/tests/IO/TestMesh.h
@@ -11,7 +11,7 @@
 
 #include "NoWarning/tut.h"
 
-#include "MeshFactory.h"
+#include "MeshDetect.h"
 #include "Reorder.h"
 #include "DerivedData.h"
 #include "ProcessControl.h"
